requirements.txt
streamlit
sqlalchemy
pymysql
langchain
python-dotenv
pandas
requests
